chore: sync 로직 분리
