+++ frontend/src/pages/Room/hooks/index.ts
@@ -1,4 +1,5 @@
 export { default as useGameHandler } from './useGameHandler';
 export { default as useGameStatus } from './useGameStatus';
+export { default as useGameSync } from './useGameSync';
 export { default as useMyTurn } from './useMyTurn';
 export { default as usePainterInfo } from './usePainterInfo';
